fixed gridlines so they line up in ggpairs
In order to fix the gridlines so they line up in ggpairs, I
manually fixed the limits of numeric variables to one percent of
the range of the variable below the min and above the max. This
negates the differences in limits that existed between the geoms. I
also removed the ticks on the plots when the user sets removeTicks
equal to true.

All that remains to be done now is to create plots along the margins
that house the labels for the axes. At this point, we can completely
remove the removeTicks option, and set those options as defaults.