add error checking for args in ggparcoord
I added some basic checks to check the arguments of ggparcoord for
validity before the function is run (to save time for the users).
I'll need to go back through and make sure these all work correctly,
but after basic testing, it seems fine.

I also fixed a bug in the ordering functionality that blew up when
a user passed a numeric vector of indices.