add ability to order axes by skewness in ggparcoord
I wrote a function to calculate the sample skewness of a vector
called skewness and added the ability to order the axes of the
parallel coordinate plot by skewness (from most skewed to least
skewed). At this time, ggparcoord just considers the absolute value
of the sample skewness, so it does not discern between left and
right skewed variables. Perhaps in the future functionality could
be added to flip variables so all are skewed in the same direction?