add ability to order axes by overall F stat in ggparcoord
Can now order the axes in ggparcoord by the overall F statistic
from an ANOVA with the class as the explanatory. In order to use
this option, the user must specify a groupColumn (class var.) to
color by. The intent of this option is to order the variables by
their variation between classes (relative to variation within
classes).

I also made a slight change to the data that ordering calculations
are being made on. Instead of the original dataset, calculations
are now being made on the data after factor/character variables
have been converted to numeric.