add ability to order axes by class separation to ggparcoord
I added another option for the order parameter. This one (currently
called classSep), instead of looking at the overall F-statistic from
an ANOVA with class as the explanatory variable, does an ANOVA for
each individual class (versus the rest) for each axis variable. The
axes are then ordered by the maximum F-statistic associated with it.
The intent is to focus more on separation between any one class and
the rest, instead of general separation between classes (since the
goal of parallel coordinate plots and ordering axes is typically
to distinguish between classes).