Change ggally_text background
The background for ggally_text had previously been a geom_rect
object, and I needed to change this slightly to be able to use
ggally_text for the internal axis labeling scheme that's upcoming.
The background is now set to white (theme_blank()), and I set the
major gridlines to a light grey.

At some point it would be best to go back and adjust how ggally_cor
uses ggally_text to see if we can get the gridlines to match up with
the rest of the plots. This can't be done presently because
print.ggpairs uses the data stored inside the plot to adjust the
range of the plot, and there is no data stored in a ggally_cor plot.