add ggally_diagAxis for internal axis labeling in ggpairs
This is a new plotting function that will be used to plot internal
axis labels along the diagonal in ggpairs, when the argument,
axisLabels == "internal". ggally_diagAxis uses ggally_text as a
base, and then plots the labels of the major gridlines at the
intersections (for continuous variables). I retrieved the axis tick
labels through the ggplotGrob and getGrob functions -- not a
particularly satisfying solution. There may be a better way to do
this in the future; one that is more efficient and does not open
a viewport.

For categorical variables, I used the levels of the variable, and
plotted them along gridlines that I calculated to roughly line up
with the facets/bars in surrounding plots.