Added internal axis labeling to ggpairs
I changed the ggpairs and print.ggpairs functions to accomodate
internal axis labeling. I added the axisLabels argument to ggpairs,
so the user can now specify the type of axis labels they want. "none"
corresponds to no labels (with variable names on the margins),
"internal" corresponds to the new internal axis labels made using
ggally_diagAxis, and "external" will correspond to axis labels on the
margins. NOTE: right now, the only functionality is to set the labels
as "internal" (which works) or "none" (actually, anything entered for
axisLabels except for "internal" should default to this). An external
axis labeling scheme should be installed in the future when there are
more flexible axis options in ggplot2. Actually...I guess I should
remove that "external" option from the documentation...

Also, the print.ggpairs function was changed to extend the plotting
region out when using the "internal" axis labeling, as variable
names are no longer needed on the margins.