change default text color in ggally_text to grey50
This changes the default color to match the text colors in the
diagonal axis plots. I originally tried to only apply this change
to ggally_cor, but couldn't get it to work. This might actually be
an okay solution, though, as it will add cohesiveness with uniform
text coloring (as a default). ggally_text is primarily an internal
function, anyway.