fix ggally_cor to remove rows appropriately
The problem I previously thought was due to improper handling of
character variables was actually due to ggally_cor improperly
removing all rows that aren't complete cases in the data set, as
opposed to only removing rows with missing values for the variables
in that particular plot. This is fixed now.